refactor(web): split RecordDetailPage.tsx into feature folder (#370)
Split the 1152-line monolithic RecordDetailPage into 13 focused modules
under apps/web/src/features/record-detail/, following the same pattern
used for FieldBuilderPage (#369). No file exceeds 400 lines.

Structure:
- types.ts — shared type interfaces
- helpers.ts — getInitials, groupFieldsBySection, sectionsFromPageLayout
- hooks/useRecord.ts — data loading (record, object def, layout)
- hooks/useRecordEdit.ts — edit form state and handlers
- hooks/useRecordDelete.ts — delete confirmation state and handlers
- hooks/useLeadConversion.ts — lead conversion state and handlers
- components/RecordHeader.tsx — breadcrumb, title, action buttons
- components/RecordFieldSection.tsx — view and edit field rendering
- components/RecordMetaFooter.tsx — owner, last modified, created
- components/RelatedRecordsList.tsx — related records with inline forms
- components/DeleteConfirmModal.tsx — delete confirmation dialog
- RecordDetailPage.tsx — orchestrator (367 lines)
- index.ts — barrel export

All 619 tests pass, typecheck clean, behavior unchanged.
