+++ apps/web/src/features/record-detail/index.ts
@@ -1,0 +1,1 @@
+export { RecordDetailPage } from './RecordDetailPage.js';
